add: multiple maps extendability
