+++ src/main/java/net/tiagofar78/prisonescape/items/SelectNoneTeamItem.java
@@ -29,9 +29,9 @@ public class SelectNoneTeamItem extends FunctionalItem {
     }
 
     @Override
-    public ItemStack toItemStack(PEPlayer player) {
+    public ItemStack toItemStack(PEGame game, PEPlayer player) {
         WaitingPlayer waitingPlayer = (WaitingPlayer) player;
-        ItemStack item = super.toItemStack(player);
+        ItemStack item = super.toItemStack(game, player);
 
         if (waitingPlayer.getPreference() == TeamPreference.RANDOM) {
             item.addUnsafeEnchantment(Enchantment.DURABILITY, 1);
