+++ src/main/java/net/tiagofar78/prisonescape/items/SelectPoliceTeamItem.java
@@ -29,9 +29,9 @@ public class SelectPoliceTeamItem extends FunctionalItem {
     }
 
     @Override
-    public ItemStack toItemStack(PEPlayer player) {
+    public ItemStack toItemStack(PEGame game, PEPlayer player) {
         WaitingPlayer waitingPlayer = (WaitingPlayer) player;
-        ItemStack item = super.toItemStack(player);
+        ItemStack item = super.toItemStack(game, player);
 
         if (waitingPlayer.getPreference() == TeamPreference.POLICE) {
             item.addUnsafeEnchantment(Enchantment.DURABILITY, 1);
